Added test for MPInteger and fixed bugs

+++ src/main/java/ch/dissem/msgpack/types/Utils.java
@@ -101,7 +101,7 @@ public class Utils {
         return new MPFloat(value);
     }
 
-    public static MPInteger mp(int value) {
+    public static MPInteger mp(long value) {
         return new MPInteger(value);
     }
